Export owner management methods
